사장님 커뮤니티 이후성 차트 페이지 한곳으로 모음(boss_chart.jsp) 신규고객과 재방문고객의 주문 금액과 주문 수 추가

+++ MAT_ZIP/MAT_ZIP/src/main/java/com/mat/zip/boss/ComController.java
@@ -1,20 +1,21 @@
 package com.mat.zip.boss;
 
-import java.util.ArrayList;
 import java.util.List;
+
+import javax.servlet.http.HttpSession;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.ResponseBody;
 
-@Controller // 스프링에서 제어하는 역할로 등록!
+@Controller
 public class ComController {
 
 	@Autowired
 	ComDAO dao;
 
+	// 댓글 등록
 	@RequestMapping("/boss/Board_insertcom")
 	public void insert(ComVO bag, int board_id, Model model) {
 		List<ComVO> list = dao.list(board_id);
@@ -24,19 +25,36 @@ public class ComController {
 		dao.insert(bag);
 	}
 
+	// 댓글 수정
 	@RequestMapping("/boss/Com_update")
-	public void update(ComVO bag) {
-		System.out.println("update요청됨.");
-		System.out.println(bag);
-		dao.update(bag);
+	public void update(ComVO bag, HttpSession session) {
+		String currentUserId = (String) session.getAttribute("user_id");
+		ComVO existingComment = dao.one(bag.getReply_id());
+		if (currentUserId.equals(existingComment.getWriter())) {
+			System.out.println("update요청됨.");
+			System.out.println(bag);
+			dao.update(bag);
+		} else {
+			// 현재 사용자가 댓글 작성자와 일치하지 않는 경우 예외를 발생시킵니다.
+			throw new RuntimeException("Only the original author can edit the comment.");
+		}
 	}
 
 	@RequestMapping("/boss/Com_delete")
-	public void delete(int reply_id) {
-		System.out.println("delete요청됨.");
-		System.out.println(reply_id);
-		dao.delete(reply_id);
+	public void delete(int reply_id, HttpSession session) {
+		String currentUserId = (String) session.getAttribute("user_id");
+		ComVO existingComment = dao.one(reply_id);
+		if (currentUserId.equals(existingComment.getWriter())) {
+			System.out.println("delete요청됨.");
+			System.out.println(reply_id);
+			dao.delete(reply_id);
+		} else {
+			// 현재 사용자가 댓글 작성자와 일치하지 않는 경우 예외를 발생시킵니다.
+			throw new RuntimeException("Only the original author can delete the comment.");
+		}
 	}
+
+
 
 	@RequestMapping("Com_one")
 	public void one(int Com_id, Model model) {
